// functions/src/models/addressField.ts
/**
 * Users API
 * Users API for the RecruitBuddy project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface AddressField { 
    addressFieldID?: string;
    addressFieldIndex?: number;
    addressFieldLabel?: string;
    addressFieldDescription?: string;
    addressFieldValue?: string;
}